modify the link problem temporry remove the check for the file access

+++ fdbuf.c
@@ -37,10 +37,10 @@ int bufload(time_fd* tf,size_t n){
      tf->tail_buf->buffer+MAXBUF)// if buf is full, add a buf
     add_buf(tf);
   //check how many space left
-  int spare=MAXBUF - (tf->tail_buf->bufptr_end 
+  size_t spare=MAXBUF - (tf->tail_buf->bufptr_end 
                       - tf->tail_buf->buffer);
   //min, decide how many can we read
-  int toread=spare;
+  size_t toread=spare;
   if(toread>n) toread=n;
   //see how many we actually read
   int cnt=read(tf->fd, tf->tail_buf->bufptr_end, toread);
@@ -56,7 +56,7 @@ int bufload(time_fd* tf,size_t n){
 
 int bufread(time_fd* src_tf, char* dst_buf  ,size_t n){
   //how many left for the current buffer
-  int toread=src_tf->header_buf->bufptr_end-
+  size_t toread=src_tf->header_buf->bufptr_end-
               src_tf->header_buf->bufptr_start;
   if(toread==0){//if current buffer is empty,remove it
     fd_buf* bf=src_tf->header_buf;
